AgenticAI_Proj1/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv
requests
pydantic
streamlit
pandas
